Refactoring: moved Fetch functions and settings outside
