+++ config.py
@@ -2,5 +2,5 @@ rooturl = "http://mp3tales.info/tales/"
 fetched_path = "../fetched/"
 downloads_path = "../files/"
 database = "../tales.db"
-first = 3
-last = 4
+first = 1
+last = 2
